Would need pagination fix
Currently separated out the loop into two respective divs, using two wp-queries. 
offset is overwritten with the use of posts_per_page, so I would need to just add a conditional with if index === 0 in the "remainder" section. 

Going to try and start / stop loop and only use one wp_query next. Reduce number of calls to database is the main goal